Add check to look for existing petprep_hmc_wf directory
This is added to accomodate cases where an error appears because there are inconsistencies between the number of sessions between subjects. The workflow runs correctly for all existing data, but fails for subjects that don't have specific sessions. Therefore, the entire command can be re-run again to obtain the final output structure.

Note: should open an issue to refactor workflow to run only on existing PET data.